Refactor gem with up-to-date structure
The gem build process is quite outdated and hard to debug.

I've run `bundle gem example --ext=c` and copied the Rakefile, Gemfile, and other files. Other changes come also from it.

The other bigger change is the organization of the C files. In order to have it work with Rake::ExtensionTask I've moved them to ext/odbc/.

I'm still not sure how to handle the odbc/utf8.

The other change is on the C extension to allow the ODBC namespace to be overwritten in order to introduce some Ruby code like ODBC::VERSION.